cassandra-stress: Provide a bogus JMX port
Scylla currently open her JMX port to the planet.
This lead cassandra stress to succeed to connect in JMX
and then to fail to reconnect when the node is down.

Provide a bogus JMX port so the JMX connection will fail
at cassandra-stress startup and will not try to reconnect.

Aleviate scylla issue #701.

Signed-off-by: Benoît Canet <benoit@scylladb.com>