sdcm.nemesis: Make ChaosMonkey thread safer
On long runs, it's likely that we'll have problems in
one random execution of disruption functions. When that
does happen, we want the test to carry on, so that we
don't get our long runs wrecked due to one bug in scylla
or the Nemesis code. Let's catch the exception and print
an error message.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>