timeout: Get rid of timeout and sleep
Some operation (SSH connexions, rpm installations) will
timeout depending of EC2 weather.

This is a cause of perpetual annoyance when you run tests
on big clusters which have impredictable startup latencies
and disk io latencies on small instances.
For example we have a 40 nodes EC2 tests that work better
on sunday.

So get rid of most of the timeouts and let the test run
his whole duration until failure.

In a futher patch we will grep the log to emit a warning
when an operation takes too much time.
This way the tests are functionals and independant of
EC2 state still we can check operations delays.

Suggested-by: Shlomi Livne <shlomi@scylladb.com>
Signed-off-by: Benoît Canet <benoit@scylladb.com>