sdcm.cluster.LoaderSet.wait_for_init: Parallel implementation
One of the good side effects of our new Remote implementation
is that now we are thread safe, meaning our limitation
of a single loader Node is gone.

For the loader nodes, since we are currently installing
the scylla tools in it and that process takes time,
running that in parallel on all nodes saves a great
deal of time. So let's introduce a parallel implementation
of .wait_for_init().

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>