sdcm: Remove timeouts from command execution
Some operation (SSH connexions, rpm installations) will
timeout depending of EC2 weather.

This is a cause of perpetual annoyance when you run tests
on big clusters which have impredictable startup latencies
and disk io latencies on small instances. For example we
have a 40 nodes EC2 tests that works better on sunday.

So get rid of most of the timeouts and let the operations
run indefinitely until the main test times out.

Signed-off-by: Benoît Canet <benoit@scylladb.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>