sdcm: Use the logging module
I was experimenting with the .format function of strings
when I wrote the first prototypes of this test suite.
Prototypes become real projects, but then we start neglecting
good practices and keep using prints.

So let's use the logging module and remove print calls.
This helps with future python 3 compatibility and makes
the code a bit more readable. Basically a new logging adapter
is being used to append contextual information of the objects
where actions are being executed on.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>