sdcm.remote: Allow passing a log file to remote commands
Sometimes it's more convenient to have a log file associated
with a given ssh command output. This way we have all the
output generated until a certain point, even if we have
unrecoverable failures.

So let's allow to pass a log file as an output file to
commands executed over ssh.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>