sdcm.tester: Add support to CQL queries
Introduce support to CQL queries. The implementation was
heavily based on cassandra-dtest, I only adapted it to
sdcm and added a few touches specific to this test suite,
and changed the method names to make them have a common
prefix. This commit also adds an example test showing how
the CQL session functionality works.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>