sdcm.cluster.Node: Refactor _wait_instance_up()
For some nemesis that depend on changing AWS instance state,
such as stop/starting instances, we realize that calling
waiting methods on those is subject to the same racing
conditions we have when a Node is started. So refactor
the code that safely waits for instance up, and allow
it to be used in other occasions.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>