sdcm.remote: Add watch_stdout_pattern
In order to determine when the c-s benchmark actually starts
for more accurate nemesis timing operations timestamps, introduce
the watch_stdout_pattern to the remote execution code. Basically
if you pass watch_stdout_pattern, when that pattern first occurs
on an SSH command output, we'll store the time since epoch for
that occurrence.

With feature, this we'll implement placement of nemesis
events on longevity test c-s graphs.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>