sdcm: Place nemesis events on c-s graphs
In order to have an idea of the effect of a nemesis on
the results of c-s, we can place the start/end events
as vertical axis on c-s graphs. Let's add one blue v-axis
on nemesis operation start and one red v-axis on nemesis
operation end by incrementing the operation log with
start/end time since epoch, and add that to the graphs
by subtracting the time c-s actually started.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>