grow_cluster_test.py: Implement grow cluster as a Nemesis class
Nemesis classes have utility structures that lets us to record
events and time them, so that we can place events into c-s generated
graphs. So let's move the add node code to a new nemesis, specific
to grow_cluster_test.py. One may argue that adding a new node is
not a destructive force so it's not a Nemesis, but surely we can
stretch that definition a little bit by including 'things that do
change the cluster layout'.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>