sdcm.cluster: Stop services on loader nodes
We use the same scylla AMI under test as a loader node,
as this simplifies things such as ensuring we'll have
an appropriate c-s version to use on said loader node.

However, we are not stopping scylla services on loader
nodes, which is likely to have an impact on the loader
process, as it's going to be competing with scylla server.

Let's pass --stop-services to the loader AMI create_instances
request to avoid having c-s and scylla server competing
out of necessity.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>