sdcm: introduce get_stress_result()
We uses cassandra-stress as a workloads, and we only check if it executes
successfully, but not parses the performance numbers.

Before introducing a new regression test that uses cassandra-stress
infrastructure in the framework, this patch introduced a help function
to collect results of all nodes and return a dictionaries' list.
Then we can easy to parse, compare, display or save it.

Signed-off-by: Amos Kong <amos@scylladb.com>