sdcm.remote: Introduce SSH key install feature
On libvirt we do not have the convenience that the
credentials service on AWS provides us. So when using
libvirt, we have to install the user's SSH key into
the new remote system. So let's use ssh-copy-id to
do that. This adds a dependency on aexpect, a library
similar to expect/pexpect, that is maintained by
yours truly and is available on pypi.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>