Add custom_cs_test.py
This is a quick and simple to modify test for people
seeking to automate running a customized cassandra-stress
command, using a YAML configuration file. Details on
how to construct and create such a file can be found on:

http://www.datastax.com/dev/blog/improved-cassandra-2-1-stress-tool-benchmark-any-schema

The included example tries to test at the same time
wide partition support and DateTieredCompactionStrategy
support. Each user may want to customize the file to
fulfill their needs.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>