sdcm.cluster: Watch out scylla-ami-setup service status as well
We recently hit a problem where one of the nodes couldn't
finish setup because it didn't get a response from the reflector.
Let's also keep an eye on scylla-ami-setup logs as part of the
db logs, so that we can quickly spot such problems without
having to actually log into the nodes.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>