sdcm.cluster: libvirt - Setup seed node first, then other nodes
Since on libvirt we are doing all the work ourselves,
when starting all nodes in parallel we are risking regular
nodes to finish booting up before the seed node. This makes
scylla to fail to start on those nodes, making the tests
hang forever. Let's first set up the seed, then the other
nodes can be safely set up in parallel.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>