sdcm: Simplify cassandra-stress interface
Instead of adding another layer of parameters that will
be later translated into a c-s command, that needs to be
constantly updated, let people simply state their c-s
command and let the framework do the rest.

Of course, the -node flag can be safely omitted, since
the address information is dynamic and we can fill that
out without a lot of trouble. In more specialized tests,
the test writer is free to fill -node, and if he/she does
that, we won't fill out -node.

This patch simplifies c-s handling in s-c-t quite a bit,
and makes it working with the suite a lot easier.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>