New test: build_cluster_test.py
Not strictly a test, but rather a way to build a cluster
in an automated fashion, with monitoring already enabled,
that you can later use to run manual testing on. One has
to set 'failure_post_behavior: destroy' in order to be
able to actually used the cluster that was created.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>