sdcm.cluster: Fix node indexing
Currently, if we remove nodes and then add new ones,
the node index won't be correct, because we are evaluating
the index based on the amount of existing nodes.

Let's keep increasing node indexes, no matter how many
nodes are destroyed and then added. This will make it
easier to analyze s-c-t logs.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>