sdcm: Allow Nemesis to reconfigure prometheus monitoring
Some nemesis might destroy and/or add new nodes into the
cluster, which requires that prometheus configuration
is updated so that we can get accurate information on
grafana dashboards.

This patch makes nemesis classes aware of the monitoring
set. This way whenever a node removal or addition is performed,
the monitoring system can be updated to watch the current
list of database cluster nodes.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>