AWSCluster: Add the mechanic to prepopulate the database
With this patch a list of EBS can be used to stop the database
after initial boot them copy the EBS contents then restart the
database.

This way tests like the grow cluster test can be run on large
dataset in a reproductible way without having to wait for hours.

Signed-off-by: Benoît Canet <benoit@scylladb.com>