sdcm: Make IP to access machine instances configurable
When one runs SCT from a cloud provider instance, in
the same internal network as the machine instances that
will be accessible, it's always advantageous to use
internal IPs, but when a work laptop is used, you have
to use the public IPs to access them. Make that aspect
configurable, and make the default value 'public'.

v2: update hostname in BaseNode

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>
Signed-off-by: Amos Kong <amos@scylladb.com>