refresh_test: optimize test time
In longevity test, persistent heavy stress is in background, we can't stop it.
But the uncompress is effected when node is serving workload, but we need stress
during the refresh.

So we can reserve current refresh monkey for longevity test, and optimize refresh
test. Start stress when sstable is uncompressed.

Signed-off-by: Amos Kong <amos@scylladb.com>