enhance to get a free port for prometheus API server
Currently we randomly choose a port to start prometheus API server,
it has little chance to fail, and it only try once.

This patch tried to start server from default 9389, it's occuried
by us in prometheus wiki. It will try to find a free port from a
range if default port isn't available.

https://github.com/prometheus/prometheus/wiki/Default-port-allocations

Signed-off-by: Amos Kong <amos@scylladb.com>