Test in mgmt_cli_test that tests the repair ability.
The test adds two keyspaces to the cluster: one with SimpleStrategy and one with LocalStrategy.
The test makes sure that all of the relevant system tables and the one with simplestrategy appear in the repair, and that the one iwith the localstrategy does not.
I also added a function in ClusterTester that grants the ability to create a generic keyspace and validates that it exists across the cluster

TO SQUASH