Changing longevity 1tb back to CL=QUORUM
Currently, with the recent changes this test is broken.
nemesis_during_prepare is true by default and prepare with CL=ALL will fail.
I'm reverting the prepare stage to be QUORUM as it was in the past and changed to catch a suspect data loss issue.
The reads are in cl=QUORUM, so it shouldn't have any issue.

(cherry picked from commit ec3304b272e80f0bbff14ac42b911b7b09f4a4f7)