refresh nemesis: correct the md5
It's very slow and unstable to download 100G files from S3 to GCE instances,
currently we actually uploaded a small file (3.6 K) to S3.

We had a solution to save the file in GCE image, it requires bigger boot disk.
But in my old test, the instance init is easy to fail. We can try to use a
split shared disk to save the 100GB file.

This patch changed the md5 to match with the file in S3.

Signed-off-by: Amos Kong <amos@scylladb.com>