Fix refresh nemesis: prepare test ks/table if it doesn't exist (#1254)
* refresh nemesis: correct the md5

It's very slow and unstable to download 100G files from S3 to GCE instances,
currently we actually uploaded a small file (3.6 K) to S3.

We had a solution to save the file in GCE image, it requires bigger boot disk.
But in my old test, the instance init is easy to fail. We can try to use a
split shared disk to save the 100GB file.

This patch changed the md5 to match with the file in S3.

Signed-off-by: Amos Kong <amos@scylladb.com>

* common/utils: remote_get_file: re-enable md5 checking

We had correct the md5 of sstable file in refresh nemesis,
let's re-enable the md5 checking in remote_get_file().

Signed-off-by: Amos Kong <amos@scylladb.com>

* nemesis: introduce _prepare_test_table

Introduce a helper to prepare test keyspace/table by simple cassandra-stress.

Signed-off-by: Amos Kong <amos@scylladb.com>

* refresh nemesis: prepare keyspace1.standard1 if it does not exist

In some longevity, the workloads won't create keyspace1.standard1. But it's
required by refresh nemesis, the backported sstable files can only be restored
to keyspace1.standard1.

Signed-off-by: Amos Kong <amos@scylladb.com>