feature(nemesis): introduced multiple StreamingError nemesises
This patch introduced some nemesises, which trigger streaming by some nodetool
task, and try to stop the task in middle to trigger some streaming fails, then
recover the node by rebuild.

This patch uses decommission, rebuild and repair to trigger streaming.

Note: In latest Scylla, it is optional to use streaming for rebuild and
decommission, and repair will not use streaming.

Signed-off-by: Amos Kong <amos@scylladb.com>