Fix issue with mising dashboards for new scylla versions
For branch-3.2 there is no appropriate dashboards in scylla-monitoring
repo. So sct installs the dashboards for master. This cause and error
that dashboard doesn't exists and screenshots and snapshots doesn't
collected correctly.

Fix added the check, that appropriate dashboard exists in Grafana,
and if not, use 'master' instead of scylla version for dashboard.

Also, code for geting correct screenshot resolution was changed