Introducing `GrafanaEventAggragator` thread
This thread checks every few minute for duplication of events in grafana
and delete some of the similar event in that time window.

this is to allow us to clearly see what's going on in the granfa more clearly.

NOTE: those events are not removed from any of the event log files