Added a get_results to the c-s on backup tests
because we weren't waiting for the results
some backup tasks started even before the c-s
created the schema, so the only backed up
schemas were the system ones.
when running the verification, the keyspace1
wasn't in there, then the sanity was failing
after downloading the files uploaded by the
backup task.
Now the test waits for c-s results, and only
then creates a backup task.