Replace the current version tagging of resources
Till now the version tag was based on GIT_BRANCH which created a mess when people used their private branches.
We would like to have it better tracked in our AWS reports and track a version cost.
In order to do that, we will use the root dir of job name to identify which job it's related to.
For example,
JOB_NAME=scylla-master/longevity/longevity-50gb-4days
The job will be tagged with version scylla-master.
In case of a missing JOB_NAME (can happen when one run test locally and doesn't pass the JOB_NAME param), it'll be tagged as "unknown".