fix(alternator): added chaos monkey nemesis and 2 more loaders
was decided that we needed to increase from 1 to 3 loaders, to generate
enough load.
also, there was no nemesis, so was decided to add some nemesis, and for
last run, there were no issues with nemesis (the ones that ran), so if
any nemesis has problems, we will either skip it (if alternator then
skip) or to remove it from alternator nemesis list.