improvement(backup nemesis): improved the success verfication mechanism
following the review of original PR, this item was postponed
to have this nemesis running already. This current PR comes
to fix the success verification, where in case of failure,
we had in the previous PR, to wait the whole timeout.

We still might want to increase the timeout, as bigger
data and metadata setups could take more than current 3 hours.