fix(large-collections): limit list items number
Currently the items number in list is unlimited, the list is too big
during the test. This patch added an update ops which will reset the
list and make it small again. In one real job, I see some 30M column,
it's acceptable. We might adjust the update ops probability or list size
in future according the feedback in reported scylla issue.

This patch also added limited set columns for large collections test.

Signed-off-by: Amos Kong <amos@scylladb.com>