fix(large-collections): use sequence population to avoid huge list
Use sequence population to avoid the repeatly insert. Repeatly insert
for list won't overwrite existing items, it will cause the list to be
very huge, the list size is random and unlimited.

For reduce the memory overload, use exponential distribution for large
collection sizes, otherwise the throughput is very low and node is easy
to be unavailable in nemesis, which might cause the cs workload fails.

There are two large collections: users_desc and names_set.

Signed-off-by: Amos Kong <amos@scylladb.com>