fix(manager agent installation): changed installation
manager-agent is now installed on a way to allow nodes
being added during a test (like on decommission or
terminate and replace) to have manager agent working.
It still a question if the manager server will know
how to deal with nodes being added and removed from
its cluster, or if extra actions will need to take
place inside the manager object.
At this point, the basic installation (on a new cluster)
works just as before, but being called from node_setup
instead of tester.py.