fix(config_setup): stop using regex to replace variables in scylla.yml
since we move to use `scylla-machine-image` the scylla.yml in AMIs is
no longer the orignal one (i.e. it the same but without comments)
so there is not point of using regexs to try to preserve it.
this fix read the yaml with pyyaml, replaces what's needed and build it back.

Trello: https://trello.com/c/jqHnwCtU