fix(mgr backup workaround): because of mermaid issue #1794
Backups are failing with timeouts when moving files, and the
suspection is that the network usage is in the limit.
Lowering the rate limit shall allow the backup to finish,
as it will demand less bandwidth.
Only a workaround, until fix is released (apparently on 2.0.2)

(cherry picked from commit d6630e7dcdca0c3db669b1d411ca6fa46012d35b)