fix(cluster health check): ignore status shutdown of node
After we replace a dead node, it can still appears in gossipstatus.
It states will be shutdown and it's ok from scylla perspective.
So node with this staus will be ignored when comparing the list of nodes
in system.peers to gossip status.
Issues:
https://github.com/scylladb/scylla/issues/6135
https://github.com/scylladb/scylla/issues/6135

(cherry picked from commit 171597052034569c39555e72490a64abe2340cae)