improvement(5k tables test): increasing test run time
the test ran before only for 50 minutes. After it took
more than 20 hours to create the whole environment,
nothing is being really tested, as within 50 minutes
(or less) there wasn't enough time to run nemesis.
We saw for example that the corrupt repair didn't even
start, as we scan all tables (and indexes) to choose
where to corrupt, and then the test finished before
we were even able to test anything.
In this first moment increasing it to 24 hours,
but we might decide later to increase it even more.