fix(wait_for_init): increase wait time between setup non-seeds nodes
Currently we setup seeds nodes firstly, the setup non-seeds nodes in
parallel. There is a 30 seconds delay between the nodes, but it's not
enough to wait the node to be ready(UP).

This patch increased the delay time to 2 mins, it's helpful to avoid
https://github.com/scylladb/scylla/issues/6354

Signed-off-by: Amos Kong <amos@scylladb.com>
(cherry picked from commit 8f122923f6644e2aa09ddc4e46a04b3107de1b22)