fix(scylla install): drop external openjdk installation
We already have strict dependency version in scylla-tools-java spec file, and
openjd-11 is already supported in (jmx, etc). So we don't need external openjdk
installation and update-java-alternatives.

This patch only changed supported distros.

On Ubuntu16.04, openjdk-8 and open-jdk-9 are available. On Ubuntu18.04,
openjdk-8 and open-jdk-11 are available. In scylla-tools-java spec file,
openjdk-8 has high priority, so it will be installed by default.

Signed-off-by: Amos Kong <amos@scylladb.com>