improvement(ipv6 workaround): remove ipv6 workaround
Amazon Linux 2 has a fix for the IPv6 persistency, hence it
doesn't need the workaround. Other AMIs still need it, so
the code is checking if not Amazon Linux 2 to run the IPv6
persistency (as it was until today).

(cherry picked from commit 96d4ab302ddc4fe16caa894551e6c70390f1fd40)