fix(rolling upgrade): remove subtest and correct schema_load_error_num
Few fixes for rolling upgrade:
1. remove using subTest
2. Issue https://github.com/scylladb/scylla-enterprise/issues/1391.
By Eliran's comment: For 'Failed to load schema version' error which is expected and non offensive
is to count the 'workload prioritization' warning and subtract that amount from the amount of
overall errors.
3. using non-frozen user type for cluster key is not supported in some versions
(https://github.com/scylladb/scylla/pull/4934). If Scylla base version doesn't support it,
ignore CQL for testing this