fix(health_check.py): reducing to error severity
because of issue scylladb/scylla-enterprise#1419.
It must be reverted once issue is fixed.
For now, when check_node_status fails, it will raise
an error event, and not a critical event, so the
execution will continue, and not halt because of the
critical event.

(cherry picked from commit ef9cc9f80efb9ebf3c1bcf873214825aff7d4aba)