fix(wait for schema agreement): wait for schema agreement on all nodes after schema change
Nemesis toggle_table_ics run alter table and finishes immediately. So the schema merge started
after the nemesis is finished and as result health checker fails on NodeSchemaVersion validation
because of schema change is running.

Another case - encription-at-rest longevity - test table is altered with scylla_encryption_options
and we need to wait for schema agreement completed