fix(longevity-in-memory): fix "Multiple definition for property 'compaction'" error
longevity-in-memory test failed with error "Multiple definition for property 'compaction'".
The issue is that there is a default value for compaction_strategy parameter.
So when the table is created, this default compaction_strategy will be defined for the table.
In case of in-memory test 'InMemoryCompactionStrategy' compaction strategy will be additionally defined.
It fails the test

(cherry picked from commit c6ffe72e178758ec82ecc2ec694d8c83fb331fc4)