fix(improve spot request): Improve spot duration request time calculation
There are the changes in this commit:
1. Use "floor" operator instead of "round" when calculate spot duration
because of "floor" round float to lower.
Example:
if we have test duration 350, we will get round(350/60)*60+60 > 360, but the maximum is 360

2. Fallback when getting spot failed.

1. change provision_type values in pipeline to:

spot
on_demand
spot_fleet

2. create a boolean YAML parameter:
instance_provision_fallback_on_demand: true|false

Fallback mechanism:

- if selected provision_type is "spot": try to get spot_duration instances
(according to test_duration)
- if test_duration > 360 or spot_duration failed, try to get spot_low_price
- if instance_provision_fallback_on_demand==true and previous attempts failed then create on_demand
 instance else fail the test

Do not change self.provision_type if fallback mechanism was used. It is OK that mixed instance
provision can happen i.e. it can happen that some of the instances in the cluster are created
as spot and other on_demand or spot duration.

If selected provision_type is "spot_fleet" - don't fallback in case failure and fail the test