fix(recognize resharding): recognize resharding in case it was finished before Scylla init completed
It happens from time to time that resharding is finished before Scylla init completed.
Because of we can run nodetool and check if the resharding has been started just after the
Scylla is up, we can't recognize that resharding was started.
This commit added searching into the log to cover this case