fix(manager): Fixed the parsing of cluster status and repair tables
Due to changes in the latest version of manager, I changed the parsing of two of the outputs
of manager: The progress table of a repair task, and the table of cluster status command.
In addition, added a wait for the manager that uses the old manager http port, for the upgrade test.