fix(longevity_5000_tables): Enable sequentially init cluster
Enable sequentially node adding to cluster with bootstrap.
Due to issue 7165, when repair_based_operations is enabled
the default behavior for scylla master and 4.2
Add node takes more time for each node to bootstrap with 5000 tables
and when we're not waiting for the other nodes to join
their bootstrap get stuck.