improvement(cdc_replication_test): truncate base tables after each round
The duration of comparison of tables between the master and replica
clusters using scylla-migrate is proportional to the size of the tables.
With 30 minute workload the comparison already takes about an hour; we
don't want it to take any longer (and it seems that scylla-migrate gives
up after some time).

Therefore we truncate the base tables between rounds so that the
comparison time is roughly the same for each round. Note that we don't
truncate the CDC log; we want it to grow and put strain on the cluster
(in particular, the TimeWindowCompactionStrategy).