feature(cdc_replication_test): collect replicator and scylla-migrate logs
We want these logs to be available for inspection after a test run.

Previously they would be printed to stdout, so they could be accessed
e.g. from the console output of a Jenkins job where we usually run this
test. However, these logs may be quite large and we don't want them to
pollute the test run logs.

Instead the logs will be available at the SCT runner and will be
collected by the "collect logs" Jenkins stage.