fix(network exception): add missed network exceptions
Two network errors are missed in the exception_retryable:
SSHConnectTimeoutError, RetryableNetworkException

NETWORK_EXCEPTIONS is defined in the sdcm.remote.init.py and collect exception_retryable
values from sdcm/remote/remote_libssh_cmd_runner.py and sdcm/remote/remote_cmd_runner.py.
It used as allowed_exceptions in retrying for get_core_pids.

The coredump thread in the test from description failed because of got SSHConnectTimeoutError.
Because of this exception is not allowed (that is wrong), the exception was raised and
failed the thread.