fix(hydra): clean know hosts when running on SCT Runner
Fixes issue: #2691.

Sometimes it happen that AWS creates a SCT Runner instance
with the same public IP address that was used for other
SCT Runner instance past. In this case Hydra won't start
on that SCT Runner since when running Docker on remote host
we can pass SSH options. Hence we need to clean the known
hosts file.