fix(scyllakillmonkey): reduce severity of event
we have seen IOExceptions raised by c-s as error
failing the whole test when a nemesis kills Scylla
process (hence the exception is expected) and the
load doesn't stop, so, lowering the severity of the
event during the specific nemesis.
Also, had to change in verify_stress_thread not to
fail the test when this IOException is found in the
loader's log, as we are already raising an event
if it is found (so this became redundant).

(cherry picked from commit 3d940d73af4588efeedb7e85e7886103d5440733)