fix(get_grafana_annotations): make the limit much larger
Now the default limit in grafana API is 100, which mean we extract only 100 of the last events
and when we use `hydra investigate show-monitor` we have last 100 events available.

In longer runs like 48h and more, this isn't helpful, and we are the context like that.

(cherry picked from commit 1625038dadc3927975467dab507acc0301c9e9b9)