improvement(sct_events): don't interpret Gemini's stderr as CRITICAL failure
If Gemini only prints some output to stderr but otherwise finishes
(with exit code != 0), we report this fact as a WARNING event instead of
CRITICAL which resulted in stopping the test.

For example, during the reboot nemesis Gemini would report (to stderr)
that it cannot connect to the rebooted node but it would still continue
running workload by sending its requests to the other nodes.

Fixes #2748.