fix(increase jmx memory): on 5k tables test jmx was crashing
due to scylladb/scylla#7609 i added more memory to the
jmx java process, and it solved the issue.
now, do we want to do this to every run, or try to
do it only for the 5000 tables (or other tests that might
be failing with the same error)?

(cherry picked from commit d87c34b777642d5b30737a7d469409a111d1ed72)