improvement(network nemesis): wait nodes UN before end of nemesis
saw a case where BlockNetwork finished, and immediately after
it ClusterHealthValidator run (and failed) because it takes
few seconds to gossip to mark the node as UN again, making the
test to fail.
Now, this nemesis will wait for all nodes to be UN before
finishing the nemesis, and allowing ClusterHealthValidator
to run on a clean cluster.