feature(configurations): add audit.yaml
This patch added a config file of audit, it will only be used for
scylla-enterprise.

The audit is only enabled for the default keyspaces and tables of
cassandra-stress and scylla-bench.

I will backport this commit to enterprise and use this config file
for some longevity jobs.

Signed-off-by: Amos Kong <amos@scylladb.com>
(cherry picked from commit 941742bac376208bd12bb975dde9addf305f4b3b)