fix(perf latency with nemesis): fix mixed workload metrics
the original work considered taking read and write metrics
from both c-s and Scylla, where in c-s there is a 3rd
metric, "mixed" for mixed workloads.
changed a bit the format of the stored data, as there
is no more meaning to have the kind of workload that
was originally introduced to deal with mixed workloads,
so now the read and write metrics from Scylla will
appear one after the other in the result table, for mixed
workloads.