feature(rolling upgrade): cover all kinds of compaction strategies
Currently LeveledCompactionStrategy and SizeTieredCompactionStrategy are ONLY
used in one subtest (table_options_test). The default SizeTieredCompactionStrategy
is used for all c-s workloads and other subtests in fill_db_data.py

This patch changed to use all kinds of compaction strategies in upgrade
test.

Backport note: IncrementalCompactionStrategy should be added when this
patch is backported to enterprise.

Signed-off-by: Amos Kong <amos@scylladb.com>