fix(gce create instance): cover destroy gce instance failure in case preemption
Failure of GCE instance creation with error 'Instance failed to start due to preemption' may happen
when spot instance has been created but was preempted. In this case the instance is stopped, not
destroyed. We check if the instance in the resources list. If the instance is still in the list -
destroy it. But we received situation when destroying was called, the instance already was removed.
So destroying failed with ResourceNotFoundError: notFound

Add 'try - except' to destoy call