improvement(cdc_replication_test): update replicator
In Scylla commit f0950e023 the format of CDC client stream tables has
changed and new clusters require an updated client to use CDC correctly.
This commit updates the link to the CDC replicator and the script that
executes it. The new replicator understands the new table and contains
additional improvements.