fix(decommission_post_action): check decommission log after checking nodetool status
Currently we have two ways to check the decommission status, one is
checking decommission log in target node, another is checking ips in
'nodetool status'. The second one is a little bit slow, decommission
state might change in the gap.

This patch moved the quick log check after checking ips in 'nodetool status'.
It didn't fix some problem we touched, but the improvement is helpful.

Signed-off-by: Amos Kong <amos@scylladb.com>