improvement(pinned severity for events): created a new severity
for events, when we want to isolate it from other
common severity, like now we are working on
reactor stalls, so it will be really easy to
dump them into a group, and manipulate it easier.
in this commit, also fixed an issue between
reactor stall events vs. backtrace (as all
reactor stalls have "Backtrace" in its text).