fix(k8s-gke): make cleanup of GKE clusters work
Cleanup of GKE clusters doesn't work now because
filtering tags we provide include "NodeType" tag which is node
specific and must not be provided for clean up of k8s clusters.
So, remove that tag from filtering to allow k8s clsuters be deleted
only based on the "test_id" tag.