improvement(k8s-monitoring): make MonitoringStatus event be warning
Make MonitoringStatus event have warning (not "error") type of severity.
As of now it is the cause for failures in lots of K8S runs where
it fails just because of having too big diff between standalone
monitoring and k8s-based one.
It is not critical now, so, make it be "warning" as of now.
It can be changed to "error" back when approach for checking monitoring
health is stabilized.