feature(sct event): add POWER_OFF event
In scylla/issues#8016, we touched a strange Power off of db nodes,
SpotTermination isn't watched, maybe it's just some unexpected power off
in aws.

This patch added a POWER_OFF event, it's helpful to address the problem.
We also ignored this event in general restart() of aws instance, restart()
of gce is implemented by soft reboot().

Signed-off-by: Amos Kong <amos@scylladb.com>