fix(k8s-logs): fix k8s log collector
All the logs we have for K8S are stored on the runner machine,
but current logic of k8s log collector assumes that we depend on
the list of nodes.
So, reuse logic from "SCT log collector" and fix set of files
to be searched in the log dir.

(cherry picked from commit 89a18f29fa3bfa084819dbc5fdcc90107e4e1f6b)