fix(mgmt agent repo): workaround to new base image
this branch is used to releases and master, so
this workaround is to enable it to work for both
cases, adding a new paramenter to the pipeline
(defaults to master support), but enabling to use
the same job to releases as well.
This until @ShlomiBalalis adds his general fix
for the manager versions/distros support.