improvement(events): delete ScyllaOperatorRestartEvent event
scylla-operator v1.2 started having 2 pod instances where one is leader
and second is replica trying to acquire lock.
And that replica may be restarted from time to time.
So, such change makes 'ScyllaOperatorRestartEvent' event become useless
and, even more, harmful, because it is going to be triggered all the
time and mark our tests as failed without real reason for it.
So, just delete it.