fix(k8s-eks): delete K8S LB service if exists recreating scylla nodes
Right now, if we try to recreate K8S scylla node we fail
on the attempt to create already existing K8S LB service which just
doesn't get deleted when it's target node/pod gets deleted.
So, consider this situation and delete such services if exist prior
to creation of new one.