fix(kubernetes): add node affinity rules for dns resolver pod
Nothing else except system and scylla pods must be scheduled on the
scylla K8S nodes.
So, add node affinity rules for dns pod resolver the way that
it will try to be scheduled on the default/auxiliary pool first and
if cannot then on a loader or monitoring ones.